requirements.txt
PyQt5
pyserial
psychopy
pyyaml
